Move mock filter impl to testing package
